HIS-33A: try moving app and query scripts to same directory
