+++ histree_backend/histree.py
@@ -1,7 +1,9 @@
 from flask import Flask, render_template
 from typing import Dict
 
-from histree_backend.data_retrieval.histree_query import HistreeQuery
+from histree_backend.histree_query import HistreeQuery
+
+
 
 
 app = Flask(__name__)
